Now with proper inits
