+++ malenov/__init__.py
@@ -1,21 +1,24 @@
-# Make initial package imports
-import segyio
-import random
-import keras
-import numpy as np
-import matplotlib.pyplot as plt
-import math
-import time
+#!/usr/bin/env python
+# -*- coding: utf 8 -*-
+"""
+### Function for n-dimensional seismic facies training /classification using Convolutional Neural Nets (CNN)
+### By: Charles Rutherford Ildstad (University of Trondheim), as part of a summer intern project in ConocoPhillips and private work
+### Contributions from Anders U. Waldeland (University of Oslo), Chris Olsen (ConocoPhillips), Doug Hakkarinen (ConocoPhillips)
+### Date: 26.10.2017
+### For: ConocoPhillips, Norway,
+### GNU V3.0 lesser license
+"""
+from . import augment
+from . import plotting
+from . import predict
+from . import segy
+from . import train
+from .malenov import master
 
-from keras.models import Sequential
-from keras.models import Model
-from keras.layers import Dense, Activation, Flatten, Dropout
-from keras.layers import Conv3D
-from keras.callbacks import EarlyStopping
-from keras.callbacks import TensorBoard
-from keras.callbacks import LearningRateScheduler
-from matplotlib import gridspec
-
-from keras.layers.normalization import BatchNormalization
-
-from shutil import copyfile
+__all__ = ["augment",
+           "malenov",
+           "plotting",
+           "predict",
+           "segy",
+           "train"
+           ]
