+++ malenov/augment/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf 8 -*-
+### ---- Functions for data augmentation ---- (Needs further development)
+
+from .randomFlip import randomFlip
+from .randomRotationXY import randomRotationXY
+from .randomRotationZ import randomRotationZ
+from .randomStretch import randomStretch
+
+__all__ = ["randomStretch",
+           "randomRotationZ",
+           "randomRotationXY",
+           "randomFlip"]
