+++ malenov/plotting/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf 8 -*-
+### ---- Functions for visualizing the predictions from the program ----
+# Make a plotting function for plotting the features
+
+from .plotNNpred import plotNNpred
+from .show_details import show_details
+from .visualization import visualization
+
+__all__ = ["visualization",
+           "show_details",
+           "plotNNpred"
+           ]
