+++ malenov/predict/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+# -*- coding: utf 8 -*-
+# Predict the output class of the given input traces
+from .cube_parse import cube_parse
+from .makeIntermediate import makeIntermediate
+from .predicting import predicting
+
+__all__ = ["predicting",
+           "makeIntermediate",
+           "cube_parse"
+           ]
