+++ malenov/segy/segy_decomp.py
@@ -1,0 +1,85 @@
+import segyio
+import numpy as np
+import matplotlib.pyplot as plt
+
+def segy_decomp(segy_file, plot_data = False, read_direc='xline', inp_res = np.float64):
+    # segy_file: filename of the segy-cube to be imported
+    # plot_data: boolean that determines if a random xline should be plotted to test the reading
+    # read_direc: which way the SEGY-cube should be read; 'xline', or 'inline'
+    # inp_res: input resolution, the formatting of the seismic cube (could be changed to 8-bit data)
+
+    # Make an empty object to hold the output data
+    print('Starting SEG-Y decompressor')
+    output = segyio.spec()
+
+    # open the segyfile and start decomposing it
+    with segyio.open(segy_file, "r" ) as segyfile:
+        # Memory map file for faster reading (especially if file is big...)
+        segyfile.mmap()
+
+        # Store some initial object attributes
+        output.inl_start = segyfile.ilines[0]
+        output.inl_end = segyfile.ilines[-1]
+        output.inl_step = segyfile.ilines[1] - segyfile.ilines[0]
+
+        output.xl_start = segyfile.xlines[0]
+        output.xl_end = segyfile.xlines[-1]
+        output.xl_step = segyfile.xlines[1] - segyfile.xlines[0]
+
+        output.t_start = int(segyfile.samples[0])
+        output.t_end = int(segyfile.samples[-1])
+        output.t_step = int(segyfile.samples[1] - segyfile.samples[0])
+
+
+        # Pre-allocate a numpy array that holds the SEGY-cube
+        output.data = np.empty((segyfile.xline.len,segyfile.iline.len,\
+                        (output.t_end - output.t_start)//output.t_step+1), dtype = np.float32)
+
+        # Read the entire cube line by line in the desired direction
+        if read_direc == 'inline':
+            # Potentially time this to find the "fast" direction
+            #start = time.time()
+            for il_index in range(segyfile.xline.len):
+                output.data[il_index,:,:] = segyfile.iline[segyfile.ilines[il_index]]
+            #end = time.time()
+            #print(end - start)
+
+        elif read_direc == 'xline':
+            # Potentially time this to find the "fast" direction
+            #start = time.time()
+            for xl_index in range(segyfile.iline.len):
+                output.data[:,xl_index,:] = segyfile.xline[segyfile.xlines[xl_index]]
+            #end = time.time()
+            #print(end - start)
+
+        elif read_direc == 'full':
+            ## NOTE: 'full' for some reason invokes float32 data
+            # Potentially time this to find the "fast" direction
+            #start = time.time()
+            output.data = segyio.tools.cube(segy_file)
+            #end = time.time()
+            #print(end - start)
+        else:
+            print('Define reading direction(read_direc) using either ''inline'', ''xline'', or ''full''')
+
+
+        # Convert the numpy array to span between -127 and 127 and convert to the desired format
+        factor = 127/np.amax(np.absolute(output.data))
+        if inp_res == np.float32:
+            output.data = (output.data*factor)
+        else:
+            output.data = (output.data*factor).astype(dtype = inp_res)
+
+        # If sepcified, plot a given x-line to test the read data
+        if plot_data:
+            # Take a given xline
+            data = output.data[:,100,:]
+            # Plot the read x-line
+            plt.imshow(data.T,interpolation="nearest", cmap="gray")
+            plt.colorbar()
+            plt.show()
+
+
+    # Return the output object
+    print('Finished using the SEG-Y decompressor')
+    return output
