+++ malenov/train/__init__.py
@@ -1,0 +1,15 @@
+#!/usr/bin/env python
+# -*- coding: utf 8 -*-
+### ---- Functions for the training part of the program ----
+# Make a function that combines the adress cubes and makes a list of class adresses
+
+from .adaptive_lr import adaptive_lr
+from .convert import convert
+from .ex_create import ex_create
+from .train_model import train_model
+
+__all__ = ["adaptive_lr",
+           "create",
+           "ex_create",
+           "train_model"
+           ]
